feat: Reorganized graphql to avoid equating data types with use cases
