--- graphql/src/adapters/graphql/initialized_adapters/__init__.py
@@ -1,3 +1,0 @@
-from src.adapters.word_generation.word_generation import UUIDWordGeneration
-
-word_generation = UUIDWordGeneration()
