+++ graphql/src/adapters/graphql/initialized_adapters/adapters.py
@@ -1,0 +1,3 @@
+from src.adapters.word_generation.word_generation import UUIDWordGeneration
+
+word_generation = UUIDWordGeneration()
